pyproject.toml
[project]
name = "thermostat-backend"
version = "0.1.0"
description = "Backend API for thermostat data management"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
    "httpx>=0.25.0",
    "asyncio-mqtt>=0.11.0",
    "python-dotenv>=1.0.0",
]
